project/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements.in -o requirements.txt
anyio==4.12.0
    # via httpx
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
exceptiongroup==1.3.1
    # via anyio
faiss-gpu==1.7.2
    # via -r ./requirements.in
filelock==3.20.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.12.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r ./requirements.in
huggingface-hub==0.36.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
lxml==6.0.2
    # via python-docx
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r ./requirements.in
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
python-docx==1.2.0
    # via -r ./requirements.in
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
regex==2025.11.3
    # via
    #   tiktoken
    #   transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   tiktoken
    #   transformers
safetensors==0.7.0
    # via transformers
scikit-learn==1.7.2
    # via sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.0
    # via -r ./requirements.in
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.12.0
    # via -r ./requirements.in
tokenizers==0.22.1
    # via transformers
torch==2.9.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   -r ./requirements.in
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.3
    # via sentence-transformers
triton==3.5.1
    # via torch
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   huggingface-hub
    #   python-docx
    #   sentence-transformers
    #   torch
urllib3==2.6.2
    # via requests
